murasaki/requirements.txt
Django>=4.2,<5.1
django-bakery>=0.13.2
django-ckeditor==6.7.0
django-parler==2.3
django-rosetta==0.10.0
factory-boy>=3.3.0
model-bakery>=1.17.0
nltk==3.8.1
pandas==2.1.4
psycopg>=3.1.14
scipy==1.11.4